fix(core): extract isSameVersion to client-safe module to prevent node:path client bundle leak
