+++ .changeset/split-version-compare-client.md
@@ -1,0 +1,5 @@
+---
+"@eventcatalog/core": patch
+---
+
+Extract isSameVersion into client-safe module to prevent node:path from leaking into client bundle
