+++ packages/core/eventcatalog/src/utils/collections/util.ts
@@ -1,6 +1,6 @@
 import type { CollectionTypes } from '@types';
 import type { CollectionEntry } from 'astro:content';
-import semver, { coerce, compare, eq, satisfies as satisfiesRange } from 'semver';
+import semver, { coerce, compare, satisfies as satisfiesRange } from 'semver';
 import path from 'node:path';
 
 // --- FILE PATH HELPERS ---
@@ -89,16 +89,7 @@ export const getVersions = (data: CollectionEntry<CollectionTypes>[]) => {
   return sortStringVersions(versions);
 };
 
-export function isSameVersion(v1: string | undefined, v2: string | undefined) {
-  const semverV1 = coerce(v1);
-  const semverV2 = coerce(v2);
-
-  if (semverV1 != null && semverV2 != null) {
-    return eq(semverV1, semverV2);
-  }
-
-  return v1 === v2;
-}
+export { isSameVersion } from './version-compare';
 
 /**
  * Sorts versioned items. Latest version first.
